ray_cluster_mandelbrot_solver/src/requirements.txt
absl-py==2.0.0
aiohttp==3.8.6
aiohttp-cors==0.7.0
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.4.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
blessed==1.20.0
blinker==1.7.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorful==0.5.5
comm==0.2.0
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==4.4.2
deepface==0.0.80
Deprecated==1.2.14
distlib==0.3.7
executing==2.0.0
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.12.4
fire==0.5.0
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.43.1
frozenlist==1.4.0
future==0.18.3
gast==0.5.4
gdown==4.7.1
google-api-core==2.12.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
googleapis-common-protos==1.61.0
gpustat==1.1.1
grpcio==1.59.0
gunicorn==21.2.0
h5py==3.10.0
idna==3.4
imageio==2.31.5
imageio-ffmpeg==0.4.9
ipykernel==6.27.1
ipython==8.16.1
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyterlab-widgets==3.0.9
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
llvmlite==0.41.0
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.7
mtcnn==0.1.1
multidict==6.0.4
nest-asyncio==1.5.8
networkx==3.2.1
numba==0.58.0
numpy==1.25.2
nvidia-ml-py==12.535.108
oauthlib==3.2.2
opencensus==0.11.3
opencensus-context==0.1.3
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
proglog==0.1.10
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-spy==0.3.14
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.13
Pygments==2.16.1
pyopencl==2023.1.4
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytools==2023.1.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
ray==2.7.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
retina-face==0.0.13
rpds-py==0.10.6
rsa==4.9
six==1.16.0
smart-open==6.4.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
tornado==6.4
tqdm==4.66.1
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
virtualenv==20.21.0
wcwidth==0.2.8
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.14.1
yarl==1.9.2
